update TextArea file structure
